Rename recurrent_network_op recurrent_op.
